Backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
google-genai==0.5.0
google-generativeai==0.8.3
psycopg[binary]==3.1.18
numpy==1.26.4
torch==2.2.2
fasttext==0.9.2
konlpy==0.6.0
